sme-ocr-service/pyproject.toml
[project]
name = "sme-ocr-service"
version = "1.0.0"
description = "Local-first OCR microservice for Semantic-Memory-Engine"
requires-python = ">=3.11"
dependencies = [
    # FastAPI & server
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    # OCR
    "paddleocr>=2.8.0",
    "paddlepaddle-gpu>=2.6.0",
    # PDF handling
    "PyMuPDF>=1.24.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
cpu = [
    "paddlepaddle>=2.6.0",  # CPU-only fallback
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
